CIV-7531 : Refactor - rename
