+++ src/main/app/client/civilServiceClient.ts
@@ -4,7 +4,8 @@ import {AssertionError} from 'assert';
 import {AppRequest} from 'common/models/AppRequest';
 import {CivilClaimResponse, ClaimFeeData} from 'common/models/civilClaimResponse';
 import {
-  ASSIGN_CLAIM_TO_DEFENDANT, CIVIL_SERVICE_AGREED_RESPONSE_DEADLINE_DATE,
+  ASSIGN_CLAIM_TO_DEFENDANT,
+  CIVIL_SERVICE_AGREED_RESPONSE_DEADLINE_DATE,
   CIVIL_SERVICE_CALCULATE_DEADLINE,
   CIVIL_SERVICE_CASES_URL,
   CIVIL_SERVICE_CLAIM_AMOUNT_URL,
